create express endpoint for team validator
